Move message memory functions to players
